// common/MISA.QLCH.CommonStores/src/models/product/IProduct.ts
export default interface IProduct {

    CodeProductList?: string;
    NameProductList?: string;
    GroupProductList?: string;
    UnitProductList?: string;
    PurchasePriceProductList?: number;
    ShowDisplayProduct?: string;
    ProductInventory?: number;
    ProductMinInventory?: number;
    ProductMaxInventory?: number;
    PriceProductList?: number;
    StatusProductList?: string;




}